added logging to all methods in author service implementation and added validation to all request body methods in rest controllers

+++ src/main/java/uz/aminasaidakhmedova/library_project/service/AuthorServiceImpl.java
@@ -14,7 +14,10 @@ import uz.aminasaidakhmedova.library_project.dto.AuthorUpdateDto;
 import uz.aminasaidakhmedova.library_project.dto.BookDto;
 import uz.aminasaidakhmedova.library_project.model.Author;
 import uz.aminasaidakhmedova.library_project.repository.AuthorRepository;
+
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 
@@ -27,21 +30,44 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public AuthorDto getAuthorById(Long id) {
-        Author author = authorRepository.findById(id).orElseThrow();
-        AuthorDto authorDto = convertEntityToDto(author);
-        return authorDto;
+        log.info("Try to find author by id {}", id);
+        Optional<Author> author = authorRepository.findById(id);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertEntityToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorDto getAuthorByName(String name) {
-        Author author = authorRepository.findAuthorByName(name).orElseThrow();
-        return convertEntityToDto(author);
+        log.info("Try to find author by name {}", name);
+        Optional<Author> author = authorRepository.findAuthorByName(name);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertEntityToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorDto getAuthorByNameBySql(String name) {
-        Author author = authorRepository.findAuthorByNameBySql(name).orElseThrow();
-        return convertEntityToDto(author);
+        log.info("Try to find author by name {}", name);
+        Optional<Author> author = authorRepository.findAuthorByNameBySql(name);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertEntityToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
@@ -52,36 +78,60 @@ public class AuthorServiceImpl implements AuthorService {
                 return criteriaBuilder.equal(root.get("name"), name);
             }
         });
-        Author author = authorRepository.findOne(specification).orElseThrow();
-        return convertEntityToDto(author);
+        log.info("Try to find author by name {}", name);
+        Optional<Author> author = authorRepository.findOne(specification);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertEntityToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorDto createAuthor(AuthorCreateDto authorCreateDto) {
+        log.info("Creating author with data: {}", authorCreateDto);
         Author author = authorRepository.save(convertDtoToEntity(authorCreateDto));
         AuthorDto authorDto = convertEntityToDto(author);
+        log.info("Created author: {}", authorDto.toString());
         return authorDto;
     }
 
     @Override
     public AuthorDto updateAuthor(AuthorUpdateDto authorUpdateDto) {
-        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow();
+        log.info("Updating author with id {}", authorUpdateDto.getId());
+        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow(() -> {
+            log.error("Author with id {} not found", authorUpdateDto.getId());
+            return new NoSuchElementException("Author not found");
+        });
         author.setName(authorUpdateDto.getName());
         author.setSurname(authorUpdateDto.getSurname());
         Author savedAuthor = authorRepository.save(author);
         AuthorDto authorDto = convertEntityToDto(savedAuthor);
+        log.info("Updated author: {}", authorDto.toString());
         return authorDto;
     }
 
     @Override
-    public void deleteAuthor(Long id){
+    public void deleteAuthor(Long id) {
+        log.info("Deleting author with id {}", id);
+        if (!authorRepository.existsById(id)) {
+            log.error("Author with id {} not found", id);
+            throw new NoSuchElementException("Author not found");
+        }
         authorRepository.deleteById(id);
+        log.info("Deleted author with id {}", id);
     }
 
     @Override
     public List<AuthorDto> getAllAuthors() {
-        List<Author> author = authorRepository.findAll();
-        return author.stream().map(this::convertEntityToDto).collect(Collectors.toList());
+        log.info("Fetching all authors");
+        List<Author> authors = authorRepository.findAll();
+        List<AuthorDto> authorDtos = authors.stream().map(this::convertEntityToDto).collect(Collectors.toList());
+        log.info("Fetched {} authors", authorDtos.size());
+        return authorDtos;
     }
 
     private Author convertDtoToEntity(AuthorCreateDto authorCreateDto) {
@@ -93,14 +143,14 @@ public class AuthorServiceImpl implements AuthorService {
 
     private AuthorDto convertEntityToDto(Author author) {
         List<BookDto> bookDtoList = null;
-        if(author.getBooks() != null) {
+        if (author.getBooks() != null) {
             bookDtoList = author.getBooks().stream()
                     .map(book -> BookDto.builder()
-                        .genre(book.getGenre().getName())
-                        .name(book.getName())
-                        .id(book.getId())
-                        .build())
-                .toList();
+                            .genre(book.getGenre().getName())
+                            .name(book.getName())
+                            .id(book.getId())
+                            .build())
+                    .toList();
         }
         AuthorDto authorDto = AuthorDto.builder()
                 .id(author.getId())
